app/requirements.txt
Django>=3.2.5,<3.3.0
mysqlclient==2.0.3
django-crispy-forms
python-decouple
gunicorn>=20.1.0,<20.2.0
django-import-export>=3.0.1,<3.1.0